chore: Add basic statistics to demo apps (#985)
## Description

Added inference times, plus for some apps TTFT, tokens/second.
Fixed T2S streaming for unfinished sentences.

### Introduces a breaking change?

- [ ] Yes
- [x] No

### Type of change

- [ ] Bug fix (change which fixes an issue)
- [ ] New feature (change which adds functionality)
- [ ] Documentation update (improves or adds clarity to existing
documentation)
- [x] Other (chores, tests, code style improvements etc.)

### Tested on

- [x] iOS
- [ ] Android

### Testing instructions

- [ ] Run all apps and check if basic statistics are correctly displayed
- [x] Check that text without end of sequence sign is processed
correctly in text to speech example

### Screenshots

<!-- Add screenshots here, if applicable -->

### Related issues

Closes #959 

### Checklist

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have updated the documentation accordingly
- [x] My changes generate no new warnings

### Additional notes

<!-- Include any additional information, assumptions, or context that
reviewers might need to understand this PR. -->

---------

+++ packages/react-native-executorch/common/rnexecutorch/models/llm/LLM.cpp
@@ -166,23 +166,23 @@ void LLM::reset() {
   runner_->reset();
 }
 
-size_t LLM::getGeneratedTokenCount() const noexcept {
+int32_t LLM::getGeneratedTokenCount() const noexcept {
   if (!runner_ || !runner_->is_loaded())
     return 0;
   return runner_->stats_.num_generated_tokens;
 }
 
-size_t LLM::getPromptTokenCount() const noexcept {
+int32_t LLM::getPromptTokenCount() const noexcept {
   if (!runner_ || !runner_->is_loaded())
     return 0;
-  return runner_->stats_.num_prompt_tokens;
+  return static_cast<int32_t>(runner_->stats_.num_prompt_tokens);
 }
 
 int32_t LLM::getVisualTokenCount() const {
   if (!runner_ || !runner_->is_loaded()) {
     return 0;
   }
-  return runner_->get_visual_token_count();
+  return static_cast<int32_t>(runner_->get_visual_token_count());
 }
 
 int32_t LLM::countTextTokens(std::string text) const {
