+++ packages/react-native-executorch/common/rnexecutorch/models/llm/LLM.h
@@ -30,8 +30,8 @@ public:
   void interrupt();
   void reset();
   void unload() noexcept;
-  size_t getGeneratedTokenCount() const noexcept;
-  size_t getPromptTokenCount() const noexcept;
+  int32_t getGeneratedTokenCount() const noexcept;
+  int32_t getPromptTokenCount() const noexcept;
   int32_t countTextTokens(std::string text) const;
   int32_t getVisualTokenCount() const;
   size_t getMemoryLowerBound() const noexcept;
